requirements.txt
spotipy
python-dotenv
musicbrainzngs
